Demo 2: Manually modifying the second file
After manually modifying the contents of `Controller.Codeunit.al`, when suggesting code for `Processor.Codeunit.al` Copilot now correctly suggests `Process` as the verb for the third procedure. This is because it has sent both `Handler` and `Controller` in the request, and has guessed that the verb must match the codeunit name.